Rename log.py -> _log.py
